2020 day9

+++ templates/cpp/@DAY@.cc
@@ -2,10 +2,10 @@
 
 #include "helpers/helpers.h"
 
-int @DAY@::puzzle1(const std::vector<std::string>& in) {
+int64_t @DAY@::puzzle1(const std::vector<std::string>& in) {
     return 0;
 }
 
-int @DAY@::puzzle2(const std::vector<std::string>& in) {
+int64_t @DAY@::puzzle2(const std::vector<std::string>& in) {
     return 0;
 }
