+++ templates/cpp/@DAY@.h
@@ -6,8 +6,8 @@
 
 class @DAY@ {
  public:
-  static int puzzle1(const std::vector<std::string>& in);
-  static int puzzle2(const std::vector<std::string>& in);
+  static int64_t puzzle1(const std::vector<std::string>& in);
+  static int64_t puzzle2(const std::vector<std::string>& in);
 
  private:
